Updated Bulk sell
[FEATURE] Items can now be sold as bulk. Synthax for adding a bulk price is /rssetprices add id:cost bulk [amount] [description]. if id is hand, the amount argument can be omitted, and the amount is set to that of handheld item. One can buy more or the same amount of bulk prices set. If I have three prices, 4 10 and 41 and I buy 12, I will pay the unitary price for 10 items (price10/10 * 12). If I take less items than the smallest bulk, I will not pay anything but I can't exit the shop. (This was not tested on custom lore or tools).

-- Code related
[FEATURE] Price class can be checked in three different modes. Similar, compatible and equals. If is similar, is the same type but amount, if is compatible, is the same type and my amount is more or equal the other's amount, if equal is the same item with same amount.

+++ src/com/github/kuben/realshopping/commands/RSPrices.java
@@ -1,15 +1,14 @@
 package com.github.kuben.realshopping.commands;
 
-import java.util.Map;
-
-import org.bukkit.ChatColor;
-import org.bukkit.Material;
-import org.bukkit.command.CommandSender;
 import com.github.kuben.realshopping.LangPack;
 import com.github.kuben.realshopping.Price;
 import com.github.kuben.realshopping.RSUtils;
 import com.github.kuben.realshopping.RealShopping;
 import com.github.kuben.realshopping.Shop;
+import java.util.Map;
+import org.bukkit.ChatColor;
+import org.bukkit.Material;
+import org.bukkit.command.CommandSender;
 
 class RSPrices extends RSCommand {
 	
@@ -95,7 +94,7 @@ class RSPrices extends RSCommand {
 	private boolean searchItem(Shop shop, Price price){
 		boolean noMatches = true;
 		if(price.getData() != -1){//Item with specific data value requested
-			int cost = shop.getPrice(price);
+			double cost = shop.getPrice(price);
 			String onSlStr = "";
 			if(shop.hasSale(price)){//There is a sale on that item.
 				int pcnt = 100 - shop.getSale(price);
